pyproject.toml
[project]
name = "ai-research-tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "fastapi>=0.117.1",
    "jupyter>=1.1.1",
    "openai>=1.109.1",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.19.0",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.2",
]
